Change the minimum manufacturer ID to 1, per the spec
See: https://tsp.esta.org/tsp/working_groups/CP/mfctrIDs.php

"Note that while ANSI E1.11, which defines the Manufacturer ID as a
16-bit value in clause D5.12, does not restrict the range, for use
with ANSI E1.20, the ID range is restricted to 0001h through 7FFFh.
The range of 8000h and above is used in ANSI E1.33 for dynamic UIDs."

For now, keeping values 8000h and above in the range. Also, note that
none of the examples will validate because they all use ID zero.